Separate application class from entry point module.
