+++ vttcompile/vttcompile.vcxproj.filters
@@ -66,6 +66,9 @@
     <ClCompile Include="..\src\Platform.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\application.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\TTAssembler.h">
